removed c_gan, moved functions into model_wrapper, updated imports
