+++ utils/model_wrapper.py
@@ -1,7 +1,59 @@
 import os
 from utils.base_wrapper import BaseModel
-from networks.c_gan import *
 import numpy as np
+import torch
+from torch.nn import init
+from networks import safa
+
+def init_weights(net, init_type='normal', init_gain=0.02):
+
+    def init_func(m):
+        classname = m.__class__.__name__
+        if hasattr(m, 'weight') and (classname.find('Conv') != -1 or classname.find('Linear') != -1):
+            if init_type == 'normal':
+                init.normal_(m.weight.data, 0.0, init_gain)
+            elif init_type == 'xavier':
+                init.xavier_normal_(m.weight.data, gain=init_gain)
+            elif init_type == 'kaiming':
+                init.kaiming_normal_(m.weight.data, a=0, mode='fan_in')
+            elif init_type == 'orthogonal':
+                init.orthogonal_(m.weight.data, gain=init_gain)
+            else:
+                raise NotImplementedError('initialization method [%s] is not implemented' % init_type)
+            if hasattr(m, 'bias') and m.bias is not None:
+                init.constant_(m.bias.data, 0.0)
+        elif classname.find('BatchNorm2d') != -1:
+            init.normal_(m.weight.data, 1.0, init_gain)
+            init.constant_(m.bias.data, 0.0)
+    net.apply(init_func)
+
+def init_net(net, init_type='normal', init_gain=0.02, gpu_ids=[]):
+    if len(gpu_ids) > 0:
+        assert (torch.cuda.is_available())
+        net.to(gpu_ids[0])
+        net = torch.nn.DataParallel(net, gpu_ids)  # multi-GPUs
+    init_weights(net, init_type, init_gain=init_gain)
+    return net
+
+
+def define_R(ret_method, polar, gpu_ids=[]):
+    net = None
+    if 'SAFA' == ret_method:
+        sa_num = 8
+        # VIGOR
+        # sate_size = (160, 160) # 320, 320
+        # pano_size = (160, 320) # 320, 640
+        # CVUSA
+        sate_size = (112, 616) #if polar else (256, 256) # this determines the output size of the spatial transformer if polar False
+        pano_size = (112, 616)
+        net = safa.SAFA(sa_num=sa_num, H1=pano_size[0], W1=pano_size[1], H2=sate_size[0], W2=sate_size[1], use_spatialtr=not polar)
+    else:
+        raise NotImplementedError('Retrieval model name [%s] is not recognized' % ret_method)
+    if len(gpu_ids) > 0:
+        assert (torch.cuda.is_available())
+        net.to(gpu_ids[0])
+        net = torch.nn.DataParallel(net, gpu_ids)
+    return net
 
 class ModelWrapper(BaseModel):
